Make KeycloakUserChecker and UserAuthValve retrieve CHE_API env var
Instead of depending on a hard-coded API endpoint for che-server,
make UserAuthValve and KeycloakUserChecker first attempt to retrieve
it from the CHE_API environment variable.

CHE_API is set to the API endpoint for che-server in workspace pods
and ensures workspaces can communicate with Che server. (e.g. when
starting a workspace in another namespace).

When CHE_API is not set, defauts to the previous, hard-coded address.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>